feat: Add support for temporary file handling during VOD uploads, integrate @gideo-llc/backblaze-b2-upload-any for enhanced B2 uploads, and update video duration handling to use file paths

+++ src/models/vod.js
@@ -77,7 +77,7 @@ const VOD = sequelize.define(
     },
     durationSeconds: {
       // Đổi tên từ duration để rõ ràng hơn là giây
-      type: DataTypes.INTEGER, // Đơn vị: giây
+      type: DataTypes.FLOAT, // Đơn vị: giây (thay đổi từ INTEGER sang FLOAT)
     },
     // createdAt và updatedAt được Sequelize quản lý tự động nếu timestamps: true
   },
